UCS654/Assignment7 - Mashup/Program_2/requirements.txt
Flask>=2.0.1
yt-dlp>=2023.3.4
moviepy>=1.0.3
pydub>=0.25.1
python-dotenv>=0.19.1
werkzeug>=2.0.0